mm, swap: free the swap cache after folio is mapped
Currently, we remove the folio from the swap cache and free the swap
cache before mapping the PTE. To reduce repeated faults due to parallel
swapins of the same PTE, change it to remove the folio from the swap
cache after it is mapped. So new faults from the swap PTE will be much
more likely to see the folio in the swap cache and wait on it.

This does not eliminate all swapin races: an ongoing swapin fault may
still see an empty swap cache. That's harmless, as the PTE is changed
before the swap cache is cleared, so it will just return and not trigger
any repeated faults. This does help to reduce the chance.

Reviewed-by: Baoquan He <bhe@redhat.com>
Signed-off-by: Kairui Song <kasong@tencent.com>